La plante pousse !
git-svn-id: file:///Users/matti/Desktop/tmp-1/flore/svn/trunk@11 0452f256-4c2a-4938-993f-bf7dc213af01

+++ src/element/plant/Plant.java
@@ -39,7 +39,7 @@ public class Plant extends Element {
         Rectangle rect = g.getClipBounds();
 
         // get the ratio for perfect plant size (1 / 3 of screen)
-        float ratio = rect.width / 3 / image.getWidth();
+        float ratio = (float)rect.width / 3 / image.getWidth();
 
         int imw = (int)(ratio * (float)image.getWidth()),
             imh = (int)(ratio * (float)image.getHeight());
